Update to latest version of sdk
* remove console warn
* move to simplifed index.ts
